[Improvement] Access the atomic variables properly + Fix compilation warnings
* In order to guarantee the atomic operations one needs
  to use atomic_read and atomic_long_read instead of accessing
  directly the counter inside the atomic_t and atomic_long_t
  so that atomic operations are guaranteed by the kernel (cross
  check with kernel drivers implementation)
* Fix compilation warnings in iccom socket if as the size_t can
  be an unsigned int or unsigned long on different platforms hence
  we cast to the biggest possible varaible and print with %lu

+++ src/iccom_socket_if.c
@@ -485,7 +485,7 @@ static int __iccom_skif_dispatch_msg_up(
 	if (data_size_bytes > ICCOM_SKIF_MAX_MESSAGE_SIZE_BYTES) {
 		iccom_skif_err("received message is bigger than max"
 				"  allowed: %lu > %d bytes; dropping;"
-				, data_size_bytes
+				, (unsigned long)data_size_bytes
 				, ICCOM_SKIF_MAX_MESSAGE_SIZE_BYTES);
 		return -ENOMEM;
 	}
@@ -498,7 +498,7 @@ static int __iccom_skif_dispatch_msg_up(
 	if (IS_ERR_OR_NULL(sk_buffer)) {
 		iccom_skif_err("could not allocate socket buffer,"
 					" req. size: %lu"
-					, NLMSG_SPACE(data_size_bytes));
+					, (unsigned long)NLMSG_SPACE(data_size_bytes));
 		return -EPIPE;
 	}
